+++ Sources/ScreenFocus/AppLinks.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+enum AppLinks {
+    static let website = URL(string: "https://therohitdas.com")!
+    static let sourceCode = URL(
+        string: "https://github.com/therohitdas/ScreenFocus"
+    )!
+    static let issues = URL(
+        string: "https://github.com/therohitdas/ScreenFocus/issues/new/choose"
+    )!
+    static let eula = URL(
+        string: "https://github.com/therohitdas/ScreenFocus/blob/main/EULA.md"
+    )!
+    static let privacy = URL(
+        string: "https://github.com/therohitdas/ScreenFocus/blob/main/docs/PRIVACY.md"
+    )!
+}