+++ Sources/ScreenFocus/ScreenFocusApp.swift
@@ -16,7 +16,6 @@ struct ScreenFocusApp: App {
 
         Settings {
             SettingsView(settings: appState.settings)
-                .frame(width: 460)
         }
     }
 }